qa_challenge/impl_1/requirements.txt
amqp==5.0.1
attrs==20.2.0
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.0.1
kombu==5.0.2
mock==4.0.2
packaging==20.4
pika==1.1.0
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
zipp==3.3.0
